add: 增加了org.voovan.http.server.RedisSessionContainer类用于启动 Redis 进行 Session 共享 imp: 将TPerformance.getProcessorCount的返回类型修改为 int

+++ Common/src/main/java/org/voovan/tools/TPerformance.java
@@ -50,7 +50,7 @@ public class TPerformance {
 	 * 获取当前系统 CPU 数
 	 * @return 系统 CPU 数
      */
-	public static double getProcessorCount(){
+	public static int getProcessorCount(){
 		return osmxb.getAvailableProcessors();
 	}
